files I changed from cs410 assignment4
These are the files I made changes from cs410 assignment4. The kNN.java
file needs to be re-organized, probably into multiple files, which is
something, I'll do on it. It currently is in the state where I was
testing the case when 100% of tags are correct. In order to change the
result  to when they are only 90% correct. I would always make some
slight changes to kNN.java