kNN implementation
The most up to date version as of August 3 2014,  7:00pm, of the kNN
with center-neighbor similarity based selection. It performs on the
selection scheme on weighted kNN and unweighted kNN and shows the
improvements in accuracy made by the center-neighbor similarity based
selection. To run, type "Python kNN.py" (without the quotes) in the
terminal. Tested and developed on Python  2.7.1, may or may not work
for different versions.